aws-account-finder/backend/requirements.txt
fastapi[all]
uvicorn 
python-dotenv
boto3
pydantic
ipaddress